requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
numpy>=1.24.0
pydicom>=2.4.0
pillow>=10.0.0
psycopg2-binary>=2.9.3
aiofiles>=23.2.1
aioredis>=2.0.0
pillow-heif>=0.10.0
